backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
motor==3.3.2
pymongo==4.6.0

# Async & Background Jobs
apscheduler==3.10.4
asyncio-contextmanager==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# OAuth2
authlib==1.3.0
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Type checking (optional but recommended)
mypy==1.7.1
